pyproject.toml
[tool.poetry]
name = "rsa_authorize"
version = "0.1.0"
description = ""
authors = ["sleepdogQQ <b8459234@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pycryptodome = "^3.15.0"
python-dotenv = "^0.20.0"
fastapi = "^0.79.0"
uvicorn = "^0.18.2"
pycryptodomex = "^3.15.0"
SQLAlchemy = "^1.4.39"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
